Remove duplicated enum
