+++ components/src/LibraryBook/Elements/UnReadDot.tsx
@@ -1,5 +1,10 @@
 import * as React from 'react';
 
+export enum READING_STATUS {
+  New = 'new',
+  Opened = 'opened',
+}
+
 export const UnReadDot: React.SFC = () => (
   <p>아직 읽지 않은 책</p>
 );
