--- components/src/LibraryBook/constants.ts
@@ -1,21 +1,0 @@
-export enum VIEW_TYPE {
-  Portrait = 'portrait',
-  Landscape = 'landscape',
-}
-
-export enum DOWNLOAD_STATUS {
-  Downloadable = 'downloadable',
-  Wating = 'wating',
-  Downloading = 'downloading',
-  Downloaded = 'downloaded',
-}
-
-export enum READING_STATUS {
-  New = 'new',
-  Opened = 'opened',
-}
-
-export enum BOOK_COUNT_UNIT {
-  Single = '권',
-  Serial = '화',
-}
